pyproject.toml
[tool.poetry]
name = "gpt-4-search"
version = "0.1.0"
description = ""
authors = ["Andy Lok <andylokandy@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gpt_4_search"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.133"
python-dotenv = "^1.0.0"
html2text = "^2020.1.16"
numpy = "^1.24.2"
tiktoken = "^0.3.3"
requests = "^2.28.2"
google-api-python-client = "^2.84.0"
openai = "^0.27.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
